Replace fallbackRef with fallbackContainer
Rather than having to manually specify at what IRI you want a
Document to be created if it does not exist yet - which often
involves a separate request to the Pod to find e.g. the location of
the storage location in the user's Profile Document -, you can now
specify in which container you want it to be created, and that
Container itself can be found in someone's Pod.