Add option to create a new nested Container
Note that this works by creating a `.dummy` file *inside* the
intended Container, which should cause the server to create the
Container on the fly. This appears to be the generally accepted way
of doing this in Solid at this point in time.
(The .dummy file is deleted afterwards in another HTTP Request.)