Make sure only relevant ACL Statements are updated
If we, for example, only look for an existing Statement that
mentions the current Agent, we might end up with a Statement that
pertains to a different Document in the same Container, or a
Statement that is restricted to a specific Origin. Thus, we have to
make sure we select for exactly the right ACL Statements.