Make fetchDocument the only exposed actor
While fetchContainer and fetchSubject are still exposed externally,
they are now marked as experimental. fetchSubjectList is not even
used internally, so it's currently marked as experimental as well,
and may even be removed completely in the future.

This means that, although you will still be able to describe
Subjects and Containers because they are important links in the
chain towards getting to a Document, the end result will always be
a Document - the relevant Subjects can then be fetched from and
added to that using the standard Tripledoc methods.

I'm not quite sure yet whether that covers the use cases, but it
clears up a bit of the confusion of when to access just the a
Subject and when to access the Document through Plandoc.
Abstracting the final step of getting from a Document to the
Subject does not feel like it's solving a real problem now, but
would be happy to hear if it does.